requirements.txt
Django==2.1.7
djangorestframework==3.9.2
requests==2.21.0
